feat(accordian): lift up state from AccordionCategory to RestaurantMenu
makes the child AccordionCategory a controlled component of the parent RestaurantMenu which itself holds the state instead of each child holding one. By default the first category is expanded(0th index). The click handler in the child uses a function showAccordianList() passed by the parent to expand itself which controls the value of another variable passed to it called showAccordianList. The child conditionally renders its list if the variable is set to true.